add sharpen

+++ Libs/Processing/include/Processing/Filters.h
@@ -16,7 +16,7 @@ namespace Filters {
     void swapChannels(Layer& image, int change);
     void emboss(Layer& image);
     void outline(Layer& image);
-    void sharpen(Layer& image, int amount);
+    void sharpen(Layer& image, float amount);
     void pixelate(Layer& image, int amount);
 
     void canny(Layer& image);
